pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "course-buddy-v2"
version = "0.1.0"
description = "SJTU Canvas replay transcript downloader"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "browser-cookie3>=0.19.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "rich>=13.0",
]

[project.scripts]
cb = "course_buddy_v2.cli:main"
cb2 = "course_buddy_v2.cli:main"
cbv2 = "course_buddy_v2.cli:main"

[tool.setuptools.packages.find]
include = ["course_buddy_v2*"]
